requirements.txt
Django==4.2.2
django-storages[google]==1.13.2
django-environ==0.10.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
google-cloud-secret-manager==2.16.1
asgiref==3.7.2
black==23.3.0
google-auth
whitenoise
structlog
# django-structlog
google-cloud-logging
